refactor(rule): extract the rule outputs component
