+++ src/style/rule.scss
@@ -62,3 +62,14 @@
     margin-right: 8px;
   }
 }
+
+.iot-form {
+  .part-header {
+    margin-bottom: 12px;
+  }
+  .sub-block-desc {
+    margin-bottom: 12px;
+    color: var(--el-text-color-secondary);
+    font-size: 12px;
+  }
+}
