+++ src/views/RuleEngine/components/IoTForm.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="iot-form">
-    <el-form label-position="top">
+    <el-form label-position="top" size="small">
       <el-card shadow="never" class="app-card">
         <div class="part-header">{{ tl('baseInfo') }}</div>
         <el-row>
@@ -17,6 +17,10 @@
 
       <el-card shadow="never" class="app-card">
         <div class="part-header">{{ tl('filterData') }}</div>
+        <div class="sub-block-desc">
+          <span>{{ tl('ruleSQLDesc') }}</span>
+          <a href="emqx.com">{{ tl('doc') }}</a>
+        </div>
         <template v-if="briefEditType">
           <el-row>
             <el-col :span="14">
@@ -93,50 +97,10 @@
           </el-row>
         </template>
       </el-card>
-
-      <el-card shadow="never" class="app-card">
-        <div class="part-header">{{ tl('output') }}</div>
-        <el-row>
-          <el-col :span="14">
-            <template v-for="(item, index) in ruleValue.outputs" :key="item">
-              <div class="outputs-item">
-                <span>
-                  <img
-                    :src="getOutputImage(item.function ? item.function : item.split(':')[0])"
-                    width="80"
-                  />
-                </span>
-                <span>
-                  <div v-if="!item.function">{{ item.split(':')[1] }}</div>
-                  <div class="output-desc">
-                    {{ (item.function ? item.function : item.split(':')[0]).toUpperCase() }}
-                  </div>
-                </span>
-                <span class="output-op">
-                  <el-button size="mini" @click="openOutputDialog(true, index)">
-                    {{ $t('Base.edit') }}
-                  </el-button>
-                  <el-button size="mini" type="danger" @click="deleteOutput(index)">
-                    {{ $t('Base.delete') }}
-                  </el-button>
-                </span>
-              </div>
-            </template>
-            <div class="outputs-item add" @click="openOutputDialog(false)">
-              <span>{{ tl('addOutput') }}</span>
-            </div>
-          </el-col>
-        </el-row>
-      </el-card>
+      <RuleOutputs v-model="ruleValue" />
     </el-form>
   </div>
   <SQLTestDialog v-model="testDialog" :test-data="testParams" :chosen-event="chosenEvent" />
-  <RuleOutputsDialog
-    v-model="showOutputDialog"
-    :output="currentOutputItem"
-    :output-disable-list="outputDisableList"
-    @submit="submitOutput"
-  />
   <SQLTemplateDrawer v-model="isShowTemplateDrawer" />
 </template>
 
@@ -163,14 +127,12 @@ import { getBridgeList, getRuleEvents } from '@/api/ruleengine'
 import { BridgeItem, RuleForm, BasicRule } from '@/types/rule'
 import { useI18n } from 'vue-i18n'
 import { cloneDeep } from 'lodash'
-import { ElMessageBox as MB } from 'element-plus'
 import parser from 'js-sql-parser'
-import { MQTTBridgeDirection, RuleOutput } from '@/types/enum'
+import { MQTTBridgeDirection } from '@/types/enum'
 import SQLTestDialog from './SQLTestDialog.vue'
-import RuleOutputsDialog from './RuleOutputsDialog.vue'
-import { OutputItem } from '@/types/rule'
 import SQLTemplateDrawer from './SQLTemplateDrawer.vue'
 import { EditPen } from '@element-plus/icons-vue'
+import RuleOutputs from './RuleOutputs.vue'
 
 const prop = defineProps({
   modelValue: {
@@ -183,18 +145,14 @@ const emit = defineEmits(['update:modelValue'])
 
 const { t } = useI18n()
 const tl = (key: string, moduleName = 'RuleEngine') => t(`${moduleName}.${key}`)
-const showOutputDialog = ref(false)
 const testDialog = ref(false)
 const bridgeList = ref([])
 const ingressBridgeList = ref([])
 const ruleEventsList = ref([])
 const outputLoading = ref(false)
 const sqlFromType = ref('topic')
-const outputDisableList: Ref<Array<string>> = ref([])
-const editIndex: Ref<number | undefined> = ref(undefined)
 const chosenEvent: Ref<RuleEvent> = ref({} as RuleEvent)
 const briefEditType = ref(true)
-const currentOutputItem: Ref<OutputItem | undefined> = ref(undefined)
 
 const isShowTemplateDrawer = ref(false)
 
@@ -276,64 +234,6 @@ const loadBridgeList = async () => {
   outputLoading.value = false
 }
 
-const calcDisableList = () => {
-  outputDisableList.value = []
-  if (!Array.isArray(ruleValue.value.outputs)) {
-    return
-  }
-  ruleValue.value.outputs?.forEach((v: OutputItem) => {
-    if (typeof v === 'string') {
-      outputDisableList.value.push(v)
-    } else if (typeof v === 'object') {
-      //republish can be duplicated
-      if (v.function === RuleOutput.Republish) return
-      v.function && outputDisableList.value.push(v.function)
-    }
-  })
-}
-
-const openOutputDialog: (edit: boolean, itemIndex?: number | undefined) => void = async (
-  edit = false,
-  itemIndex,
-) => {
-  showOutputDialog.value = true
-  let item: OutputItem | undefined
-  editIndex.value = itemIndex
-  if (itemIndex !== undefined && Array.isArray(ruleValue.value.outputs)) {
-    item = ruleValue.value.outputs?.[itemIndex]
-  }
-  if (edit) {
-    currentOutputItem.value = item
-  } else {
-    currentOutputItem.value = undefined
-  }
-  calcDisableList()
-}
-
-const submitOutput = (opObj: OutputItem) => {
-  const output = ruleValue.value.outputs || []
-  if (Array.isArray(output)) {
-    if (!currentOutputItem.value) {
-      output.push(opObj)
-    } else {
-      editIndex.value !== undefined && output.splice(editIndex.value, 1, opObj)
-    }
-  }
-  calcDisableList()
-}
-
-const deleteOutput = async (itemIndex: number | undefined) => {
-  await MB.confirm(t('Base.confirmDelete'), {
-    confirmButtonText: t('Base.confirm'),
-    cancelButtonText: t('Base.cancel'),
-    type: 'warning',
-  })
-  if (itemIndex !== undefined && Array.isArray(ruleValue.value.outputs)) {
-    ruleValue.value.outputs?.splice(itemIndex, 1)
-    calcDisableList()
-  }
-}
-
 const loadIngressBridgeList = async () => {
   await loadBridgeList()
   ingressBridgeList.value = bridgeList.value.filter(
@@ -385,15 +285,6 @@ const loadRuleEvents = async () => {
   }
 }
 
-const getOutputImage = (item: string) => {
-  try {
-    return require(`@/assets/img/${item}.png`)
-  } catch (e) {
-    //May it be a user defined module
-    console.log('ImgErr:', e)
-  }
-}
-
 onMounted(() => {
   loadIngressBridgeList()
   loadRuleEvents()
@@ -401,12 +292,13 @@ onMounted(() => {
 })
 </script>
 
+<style lang="scss">
+@import '~@/style/rule.scss';
+</style>
+
 <style lang="scss" scoped>
 .el-card {
   margin-bottom: 20px;
-  .part-header {
-    margin-bottom: 12px;
-  }
   .el-form-item {
     margin-bottom: 12px;
   }
@@ -428,57 +320,5 @@ onMounted(() => {
 .sql-ft {
   display: flex;
   justify-content: space-between;
-}
-.outputs-item {
-  height: 92px;
-  border: var(--el-border-base);
-  margin-top: 10px;
-  display: flex;
-  align-items: center;
-  padding: 10px;
-  box-sizing: border-box;
-
-  span:nth-child(2) {
-    flex-grow: 1;
-
-    div {
-      line-height: 200%;
-    }
-
-    .output-desc {
-      color: #5b5b5b;
-    }
-  }
-
-  .output-op {
-    padding: 0 10px;
-    visibility: hidden;
-  }
-
-  &.add {
-    justify-content: center;
-  }
-  &:first-of-type {
-    margin-top: 20px;
-  }
-  &:hover {
-    border-color: var(--el-color-primary);
-    cursor: pointer;
-    span {
-      color: var(--el-color-primary);
-      visibility: visible;
-    }
-  }
-}
-
-.edit-output {
-  color: var(--el-color-primary);
-  line-height: 50px;
-  cursor: pointer;
-}
-
-.embedded-config {
-  border: var(--el-border-base);
-  padding: 30px;
 }
 </style>
